test(core): split workspace federation responsibilities
(cherry picked from commit b1fe37d70dfc87a59d1df8e80f26d2c0590d0570)
